Adding new date format and moving to own file
